requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
astroid==3.3.11
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.5
black==24.10.0
bleach==6.2.0
blinker==1.9.0
brax==0.13.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.3
chex==0.1.90
click==8.2.1
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
distlib==0.4.0
dm-control==1.0.31
dm-env==1.6
dm-tree==0.1.9
etils==1.13.0
execnet==2.1.1
executing==2.2.0
fastjsonschema==2.21.2
filelock==3.19.1
flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.2.10
flax==0.11.1
fonttools==4.59.1
fqdn==1.5.1
fsspec==2025.7.0
gast==0.4.0
gitdb==4.0.12
gitpython==3.1.45
glfw==2.9.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
-e file:///home/ps/Documents/hector_dev
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
humanize==4.12.3
identify==2.6.13
idna==3.10
immutabledict==4.2.1
importlab==0.8.1
importlib-resources==6.5.2
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.4.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
isort==6.0.1
itsdangerous==2.2.0
jax==0.5.1
jax-cuda12-pjrt==0.5.1
jax-cuda12-plugin==0.5.1
jaxlib==0.5.1
jaxopt==0.8.5
jedi==0.19.2
jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter-server==2.17.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.6
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
keras==3.11.3
kiwisolver==1.4.9
labmaze==1.0.6
lark==1.2.2
lcm==1.5.1
libclang==18.1.1
libcst==1.8.2
lxml==6.0.1
markdown==3.8.2
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mediapy==1.2.4
mistune==3.1.3
ml-collections==1.1.0
ml-dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
mujoco==3.3.6.dev798860951
mujoco-mjx @ git+https://github.com/google-deepmind/mujoco@d77ecb3f064678cc9227cdb8e0f5fccad97569fc#subdirectory=mjx
mypy==1.17.1
mypy-extensions==1.1.0
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
nodeenv==1.9.1
notebook==7.4.5
notebook-shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.12.0.46
nvidia-cufft-cu12==11.4.1.4
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nccl-cu12==2.27.7
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu12==3.3.24
oauthlib==3.3.1
onnx==1.17.0
onnxruntime==1.22.1
opt-einsum==3.4.0
optax==0.2.5
optree==0.17.0
orbax-checkpoint==0.11.23
overrides==7.7.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
pip==25.2
platformdirs==4.3.8
-e file:///home/ps/Documents/hector_dev/mujoco_playground
pluggy==1.6.0
pre-commit==4.3.0
prometheus-client==0.22.1
prompt-toolkit==3.0.51
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycnite==2024.7.31
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydot==4.0.1
pygments==2.19.2
pyink==24.10.1
pylint==3.3.8
pyopengl==3.1.10
pyparsing==3.2.3
pytest==8.4.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytype==2024.10.11
pyyaml==6.0.2
pyzmq==27.0.2
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
ruff==0.12.10
scipy==1.16.1
send2trash==1.8.3
sentry-sdk==2.35.0
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardx==2.6.4
tensorflow-cpu==2.20.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorstore==0.1.76
termcolor==3.1.0
terminado==0.18.1
tf2onnx==1.16.1
tinycss2==1.4.0
toml==0.10.2
tomlkit==0.13.3
toolz==1.0.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
treescope==0.1.10
trimesh==4.7.4
types-python-dateutil==2.9.0.20250822
typing-extensions==4.14.1
typing-inspection==0.4.1
uri-template==1.3.0
urllib3==2.5.0
virtualenv==20.34.0
wandb==0.21.1
warp-lang==1.9.0.dev20250824
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.14.2
zipp==3.23.0
